mcp/requirements.txt
mcp
qdrant-client
langchain-community
python-dotenv
sentence-transformers